requirements.txt
yfinance==0.2.28
pandas==2.1.4
gspread==5.12.4
oauth2client==4.1.3
python-telegram-bot==20.7
APScheduler==3.10.4
requests==2.31.0
ta==0.10.2
google-generativeai==0.3.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
investpy==1.0.8
numpy==1.24.3
schedule==1.2.0